intermediate work and code refactoring
